+++ lua/snippets/luasnip/init.lua
@@ -1,2 +1,3 @@
 require('snippets.luasnip.go')
+require('snippets.luasnip.lua_snippets')
 require('luasnip.loaders.from_vscode').lazy_load()